### 2020-09-22 #2686 @pibion
@kyleniemeyer I'd love to edit this paper.  Telescopes are outside my area of instrumental expertise, so I'll be looking for three reviewers.  So if you know of any editors who have expertise in telescopes, feel free to boot me?
