### 2020-09-24 #2686 @pibion
@arfon thanks, that helps!  I've also got a lead from a friend in telescope-astro.  I've got a bunch of emails out :)
