### 2020-10-02 c5a62ca
put ensure_numpy in second frg plot to avoid error when running with cupy
