+++ paper.bib
@@ -30,8 +30,8 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{hippler:2019b,
-       author = {{Hippler}, Stefan and {Feldt}, Markus and {Bertram}, Thomas and {Brand
-        ner}, Wolfgang and {Cantalloube}, Faustine and {Carlomagno}, Brunella and
+       author = {{Hippler}, Stefan and {Feldt}, Markus and {Bertram}, Thomas and {Brandner}, Wolfgang
+         and {Cantalloube}, Faustine and {Carlomagno}, Brunella and
          {Absil}, Olivier and {Obereder}, Andreas and {Shatokhina}, Iuliia and
          {Stuik}, Remko},
         title = "{Single conjugate adaptive optics for the ELT instrument METIS}",
