### 2020-10-09 2c02449
rewrote logging setup...
