### 2020-10-14 #2686 @mvandam
I could possibly do it, but I am not sure what the review entails. I don't know much about JOSS! Is it a review of the paper describing the software or is it also a review of the software? Can I see the paper and the evaluation form (if any) before agreeing to review the submission? Thanks!
