### 2020-10-14 #2686 @frigaut
Yes, I have to admit I am in the same category. I am very much in favour of open journals, and would love to help, but your message lack context and a personal touch. Who is sending this email?
Cheers
Francois
