### 2020-10-15 #2686 @pibion
@mvandam , @frigaut this is Amy Roberts - I'm one of the editors of JOSS (https://joss.theoj.org/about).

The review is a review of both the software and the paper.  JOSS reviews are a bit different from traditional journal reviews in that once it's been decided that the paper is in scope, the purpose of the review is to work with the author to improve the code so that it meets the basic JOSS publication standards.

So the review will consist of going through the checklist (https://joss.readthedocs.io/en/latest/review_checklist.html) here on github.

Let me know if you have any more questions!  I'm also happy to find a time to chat if that'd help.
