### 2020-10-17 #2686 @pibion
@mvandam excellent, thanks!  There will be a bit of a wait - we need at least one more reviewer - and then I'll start the review.

At that point, I'll open a new issue and invite you (and all the other reviewers) to the issue.  "Doing the review" will consist of going through the checklist.  In some cases, checking off boxes will require some back-and-forth with the author.
