### 2020-10-18 #2686 @frigaut
Alright, I am curious about this ne wkind of review (at least new for me), and I would be interested to give it a go too, to the extent of my abilities.
Count me in.
Francois
