### 2020-10-20 #2764 @pibion
@mvandam , @frigaut, could you both make sure you've got the ability to check and uncheck your checkboxes?

You'll need to accept the invitation to this issue and also be logged in to do so.
