### 2020-10-20 #2764 @mvandam
I have read the documentation, but I am still at a loss as to how I actually submit a written review (other than checking the boxes). Am I supposed to write the review as text here?
