### 2020-10-23 0f1efe1
Hotfix fo cases where correlation length determination fails...
