### 2020-10-29 #2764 @mvandam
The paper meets the criteria set out, but is very short. I am not a regular reader of JOSS, but I would have expected to see
(1) a comprehensive list of the important features (I know that the software is evolving), (2)  an idea of how long it takes to run on standard hardware and (3) a list of user cases (other than METIS). The paper does not do a particularly good job of "selling" the software.

As far as I can tell, AOSAT does not have a wide user base (I am happy to be proven wrong!) and I do not think that AOSAT will be widely adopted in the community. There are two major reasons for this.

First, the software does not appear to do a lot. From what I can see, it takes the residual wavefront produced by another package and computes some useful metrics, such as PSF, segment piston and contrast. Computing the PSF and integrating it with time takes literally 3 or 4 lines of code. Computing segment piston as a function of iteration requires less than a dozen lines of code. The post-coronograph contrast is useful, but every user will likely have their own coronograph and their own metrics It also appears to be hard coded for on-axis optimization and a single wavelength, which again limits the applicability of the software to classical on-axis adaptive optics systems.

The second issue is the lack of a comprehensive manual. While I could see how to change the "sensing wavelength" (which I would call the science wavelength, since the sensing wavelength in adaptive optics generally refers to the wavefront sensor), there are a lot of undocumented critical parameters (e.g., ppm : 10.0). I also do not know how I would go about inserting my own coronagraph or whatever to process the data, let alone how to process multiple wavelengths or science targets, or use an extended science target and calculate the PSF.

Finally, many potential users (e.g., high contrast imaging) intend to feed the output of the science image (e.g., non-common path aberrations sensed with the science camera) back to the adaptive optics system. So having a separate post-processing tool will not meet their goals.

I think an experienced adaptive optics scientist who has never used AOSAT would find it much easier to write their own code to process the residual wavefronts than to figure out how to modify the parameter files and the AOSAT code to produce what they want.

I have not tested the accuracy of the results (e.g., by comparing the output of a different code) but I have no reason to doubt that the results are accurate.

The software installed fine in Ubuntu Linux. I had trouble installing the software locally as described, but I am almost certain that this was related to my local installation of Anaconda. Rather than trying as a different user, I installed it as root without any issues. Other people might have the same problem if they have anaconda installed, so it might be worth pointing this out.
