### 2020-10-30 #2764 @mvandam
Having the manual makes a big difference! I will read the manual and rewrite my review accordingly in the next two or three days.
