### 2020-11-01 #2764 @mvandam
Having read the manual, I can better understand the motivation for wanting to disseminate the tool, and why someone might want to learn how to use it. Please fix the following:
- the installation page on the manual is missing all of the commands (see screen shots).
![Screenshot from 2020-11-02 10-37-36](https://user-images.githubusercontent.com/622495/97816078-7c217780-1cf7-11eb-8eb1-2318442ba363.png)
Also, the
The rest of the documentation is good. Once this is done, I will pass it on the "Substantial scholarly effort" tickbox.

What is still missing is

I don't see this anywhere.
