### 2020-11-04 f2b4712
more changes to code blocks to improve rendering
