### 2020-11-04 138fd1b
Adapted installation guide
