### 2020-11-04 #2764 @mfeldt
I cannot even begin to appreciate your patience, Marcos...

And I'm truly sorry, but it appears that most code blocks were not only
missing from the installation page, but from all others as well.
Something with the code-block:: directive when no language is specified.

Additionally, I added a section on community guidelines under the
general concept section (for extending AOSAT).

So whatever the publication, it's already clearly much better now than
it was before - thank you so much!

   markus

Am 01.11.2020 um 22:43 schrieb Marcos van Dam:
>
> Having read the manual, I can better understand the motivation for
> wanting to disseminate the tool, and why someone might want to learn
> how to use it. Please fix the following:
>
>   * the installation page on the manual is missing all of the commands
>     (see screen shots).
>     Screenshot from 2020-11-02 10-37-36
>     <https://user-images.githubusercontent.com/622495/97816078-7c217780-1cf7-11eb-8eb1-2318442ba363.png>
>     Also, the
>     The rest of the documentation is good. Once this is done, I will
>     pass it on the "Substantial scholarly effort" tickbox.
>
> What is still missing is
>
> Community guidelines: Are there clear guidelines for third parties
> wishing to 1) Contribute to the software 2) Report issues or problems
> with the software 3) Seek support
>
> I don't see this anywhere.
>
