### 2020-11-04 #2764 @mvandam
The documentation displays properly now.

Please update the documentation regarding cupy-cuda110 since it **does** exist (i.e., you should check whether this works properly or not). I did not find the SVD used anywhere, so I am not sure why this is even an issue. Generally speaking, you can do regularized least squares inversions much faster than SVD inversions if this is what you are using it for.

Other than that, I am finished with this review.
