### 2020-11-18 294569f
Hotfix to the fragment piston analyzer which got the mean piston per fragment wrong due to a coordinate sytsem error
