+++ .travis.yml
@@ -1,5 +1,5 @@
 language: python
 install:
-  - pip install
+  - pip install .
 script:
   - python setup.py test
