### 2021-05-11 #2764 @andrewpaulreeves
Hi @pibion - thanks for checking in! I have installed and begun to look at the software, but I might need another couple of weeks to test it further and write the review due to some other work commitments.

A question on the process of the review - shall I prepare a full review then "submit", or just ask questions or make comments here when they occur during my testing?

Thanks!
