### 2021-05-14 906f413
Hotfix for removing special characters (like '\t') from kex names in setup files...
