### 2021-06-23 #2764 @andrewpaulreeves
Hi @pibion. Sorry, I got a bit caught up in some other tasks, but should have time again for the review in the coming weeks.

After cloning and installing, my first comments are that:
- The repo and install is very large. The repo is 2GB, and the installed size is 1.6GB The vast majority of this is data used as examples in  ``src/aotsat/examples``, so I assume isn't vital for the operation of the code itself. Could this data be hosted elsewhere as an optional download if a user would like to see an example? Currently, having a copy of the source and an installed version of the code uses nearly 4GB of disk space, which is quite a lot for a single project, even with modern drive capacities. I imagine such a large repo size could also make development quiet unweildy moving forward, as simple clones, pushes and pulls become slower.
- There are currently no tagged "versions". Version tags are useful to track in the introduction of new features and so that the maintainers can deal with issues and bug reports as they come in. It can be very daunting to release "Version 1", but perhaps the current version should be "v0.1"?
I would also recommend looking at the python library [versioneer](https://github.com/python-versioneer/python-versioneer), which will look at the latest git tag and set the ``aosat.__version__`` attribute so the version can be accessed programatically.

I'll try and take a closer look at the functionality in the coming days.
