### 2021-06-24 #2764 @pibion
@andrewpaulreeves thank you for the update!

@mfeldt I thought we had discussed the repo size but I can't seem to find the commits.  A GB checkout is indeed pretty large.  Zenodo will host data files that large for free and also can create a DOI for them.

For versioning, I'm a big fan of "semantic versioning" and always start my software at v0.0.1!  There's a nice discussion at https://semver.org/.
