### 2021-12-02 27f190b
Added TT removal in frg analyzer
