+++ src/aosat/analyzers_/frg_analyzer.py
@@ -99,21 +99,27 @@ class frg_analyzer():
     """
     def __init__(self,sd):
         num_fragments = np.max(sd['fragmask']).item()
-        self.sd      = sd
-        self.ffed    = 0
-        self.piston  = np.zeros(num_fragments)
-        self.dpiston = np.zeros(num_fragments)
-        self.pistont = None
-        self.ttx     = np.zeros(num_fragments)
-        self.dttx    = np.zeros(num_fragments)
-        self.tty     = np.zeros(num_fragments)
-        self.dtty    = np.zeros(num_fragments)
-        self.ttxt    = None
-        self.ttyt    = None
-        self.wfrag   = []
-        self.x       = None
-        self.y       = None
-        self.A       = []
+        self.sd       = sd
+        self.ffed     = 0
+        self.piston   = np.zeros(num_fragments)
+        self.dpiston  = np.zeros(num_fragments)
+        self.pistonR  = np.zeros(num_fragments)
+        self.dpistonR = np.zeros(num_fragments)
+        self.pistont  = None
+        self.pistontR = None
+        self.ttx      = np.zeros(num_fragments)
+        self.dttx     = np.zeros(num_fragments)
+        self.tty      = np.zeros(num_fragments)
+        self.dtty     = np.zeros(num_fragments)
+        self.ttxt     = None
+        self.ttyt     = None
+        self.ttxtR    = None
+        self.ttytR    = None
+        self.wfrag    = []
+        self.x        = None
+        self.y        = None
+        self.A        = []
+        self.BigA     = None
         self.pistframe = None
         self.tile      = 0.8
         self.worstid   = 0
@@ -121,40 +127,60 @@ class frg_analyzer():
         for i in range(num_fragments):
             self.wfrag.append(np.where(sd['fragmask'] == (i+1)))
 
+    def remove_tt(self,frame,wtm):
+        C,_,_,_ = np.linalg.lstsq(self.BigA, frame[wtm])
+        d_frame = (C[2]+C[0]*self.x+C[1]*self.y)*self.sd['tel_mirror']
+        #import pdb; pdb.set_trace()
+        return((frame-d_frame))
 
     def feed_frame(self,frame,nframes):
         num_fragments = np.max(self.sd['fragmask']).item()
+        wtm = np.where(self.sd['tel_mirror'] != 0)
 
         if self.ffed == 0:
             sdim = frame.shape[0]
             self.pistont = np.zeros((nframes,num_fragments))
             self.ttxt = np.zeros((nframes,num_fragments))
             self.ttyt = np.zeros((nframes,num_fragments))
+            self.pistontR = np.zeros((nframes,num_fragments))
+            self.ttxtR = np.zeros((nframes,num_fragments))
+            self.ttytR = np.zeros((nframes,num_fragments))
             self.x, self.y = np.mgrid[-sdim/2:sdim/2,-sdim/2:sdim/2]/self.sd['ppm']
             for i in range(num_fragments):
                 xc = np.mean(self.x[self.wfrag[i]])
                 yc = np.mean(self.y[self.wfrag[i]])
                 self.A.append(np.c_[self.x[self.wfrag[i]]-xc, self.y[self.wfrag[i]]-yc, self.x[self.wfrag[i]]*0.0+1.0])
+            xc = np.mean(self.x[wtm])
+            yc = np.mean(self.y[wtm])
[…]
