### 2022-08-06 #2764 @mfeldt
Actually it's a bit late now. The repository has been active for quite some time, and the paper is essentially published at SPIE 2020... I think we can forget about it here.
